+++ lib/loggerator/rails/log_subscriber.rb
